Moved replay to a blueprint and cleaned imports
